+++ system_modes/changelog.rst
@@ -4,6 +4,7 @@ Changelog for package system_modes
 
 0.3.0 (2020-07-23)
 -----------
+* removed boost dependencies (was: program options)
 * changed mode service specifications (less redundancy)
 * https://github.com/micro-ROS/system_modes/issues/24